refactor: delete 기능 수정
- 반환 타입 수정

+++ story-service/src/main/kotlin/com/beomsic/storyservice/adapter/out/persistence/StoryDeleteAdapter.kt
@@ -1,7 +1,8 @@
 package com.beomsic.storyservice.adapter.out.persistence
 
 import com.beomsic.storyservice.application.port.out.StoryDeletePort
-import com.beomsic.storyservice.infrastructure.persistence.StoryRepository
+import com.beomsic.storyservice.domain.exception.StoryNotFoundException
+import com.beomsic.storyservice.domain.exception.UnauthorizedStoryAccessException
 import org.springframework.stereotype.Component
 
 @Component
@@ -9,7 +10,9 @@ class StoryDeleteAdapter (
     private val storyRepository: StoryRepository
 ) : StoryDeletePort {
 
-    override fun deleteStory(storyId: Long) {
+    override suspend fun deleteStory(userId: Long, storyId: Long) {
+        val entity = storyRepository.findById(storyId) ?: throw StoryNotFoundException()
+        if (entity.authorId != userId) throw UnauthorizedStoryAccessException()
         storyRepository.deleteById(storyId)
     }
 }
